+++ client/graph.c
@@ -144,11 +144,14 @@ int GetAskClock(const char str[], bool printAns, bool verbose)
 			PrintAndLog("Failed to copy from graphbuffer");
 		return -1;
 	}
-	bool st = DetectST(grph, &size, &clock);
-	int start = 0;
+	//, size_t *ststart, size_t *stend
+	size_t ststart = 0, stend = 0;
+	bool st = DetectST(grph, &size, &clock, &ststart, &stend);
+	int start = stend;
 	if (st == false) {
 		start = DetectASKClock(grph, size, &clock, 20);
 	}
+	setClockGrid(clock, start);
 	// Only print this message if we're not looping something
 	if (printAns || g_debugMode) {
 		PrintAndLog("Auto-detected clock rate: %d, Best Starting Position: %d", clock, start);
@@ -197,6 +200,7 @@ int GetPskClock(const char str[], bool printAns, bool verbose)
 	size_t firstPhaseShiftLoc = 0;
 	uint8_t curPhase = 0, fc = 0;
 	clock = DetectPSKClock(grph, size, 0, &firstPhaseShiftLoc, &curPhase, &fc);
+	setClockGrid(clock, firstPhaseShiftLoc);
 	// Only print this message if we're not looping something
 	if (printAns){
 		PrintAndLog("Auto-detected clock rate: %d", clock);
@@ -223,6 +227,7 @@ uint8_t GetNrzClock(const char str[], bool printAns, bool verbose)
 	}
 	size_t clkStartIdx = 0;
 	clock = DetectNRZClock(grph, size, 0, &clkStartIdx);
+	setClockGrid(clock, clkStartIdx);
 	// Only print this message if we're not looping something
 	if (printAns){
 		PrintAndLog("Auto-detected clock rate: %d", clock);
@@ -241,10 +246,12 @@ uint8_t GetFskClock(const char str[], bool printAns, bool verbose)
 
 
 	uint8_t fc1=0, fc2=0, rf1=0;
-	uint8_t ans = fskClocks(&fc1, &fc2, &rf1, verbose);
+	int firstClockEdge = 0;
+	uint8_t ans = fskClocks(&fc1, &fc2, &rf1, verbose, &firstClockEdge);
 	if (ans == 0) return 0;
 	if ((fc1==10 && fc2==8) || (fc1==8 && fc2==5)){
 		if (printAns) PrintAndLog("Detected Field Clocks: FC/%d, FC/%d - Bit Clock: RF/%d", fc1, fc2, rf1);
+		setClockGrid(rf1, firstClockEdge);
 		return rf1;
 	}
 	if (verbose){
@@ -253,7 +260,7 @@ uint8_t GetFskClock(const char str[], bool printAns, bool verbose)
 	}
 	return 0;
 }
-uint8_t fskClocks(uint8_t *fc1, uint8_t *fc2, uint8_t *rf1, bool verbose)
+uint8_t fskClocks(uint8_t *fc1, uint8_t *fc2, uint8_t *rf1, bool verbose, int *firstClockEdge)
 {
 	uint8_t BitStream[MAX_GRAPH_TRACE_LEN]={0};
 	size_t size = getFromGraphBuf(BitStream);
@@ -265,8 +272,8 @@ uint8_t fskClocks(uint8_t *fc1, uint8_t *fc2, uint8_t *rf1, bool verbose)
 	}
 	*fc1 = (ans >> 8) & 0xFF;
 	*fc2 = ans & 0xFF;
-
-	*rf1 = detectFSKClk(BitStream, size, *fc1, *fc2);
+	//int firstClockEdge = 0;
+	*rf1 = detectFSKClk(BitStream, size, *fc1, *fc2, firstClockEdge);
 	if (*rf1==0) {
 		if (verbose || g_debugMode) PrintAndLog("DEBUG: Clock detect error");
 		return 0;