+++ client/ui.c
@@ -12,15 +12,19 @@
 #include <stdarg.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdbool.h>
 #include <readline/readline.h>
 #include <pthread.h>
 
 #include "ui.h"
 
-double CursorScaleFactor;
+double CursorScaleFactor = 1;
 int PlotGridX, PlotGridY, PlotGridXdefault= 64, PlotGridYdefault= 64, CursorCPos= 0, CursorDPos= 0;
 int offline;
 int flushAfterWrite = 0;  //buzzy
+int GridOffset = 0;
+bool GridLocked = false;
+
 extern pthread_mutex_t print_lock;
 
 static char *logfilename = "proxmark3.log";