+++ client/ui.h
@@ -11,6 +11,8 @@
 #ifndef UI_H__
 #define UI_H__
 
+#include <stdbool.h>
+
 void ShowGui(void);
 void HideGraphWindow(void);
 void ShowGraphWindow(void);
@@ -19,8 +21,9 @@ void PrintAndLog(char *fmt, ...);
 void SetLogFilename(char *fn);
 
 extern double CursorScaleFactor;
-extern int PlotGridX, PlotGridY, PlotGridXdefault, PlotGridYdefault, CursorCPos, CursorDPos;
+extern int PlotGridX, PlotGridY, PlotGridXdefault, PlotGridYdefault, CursorCPos, CursorDPos, GridOffset;
 extern int offline;
 extern int flushAfterWrite;   //buzzy
+extern bool GridLocked;
 
 #endif